fix: three long-read bugs; Rust and C++ now produce identical output
## Bug 1: C++ correct_for_gaps reads uninitialized InsExon.gexon.exon_id (UB)

When get_intervals creates an InsExon (small query exon with no guide
match), it sets has_gexon=false but leaves the Segment struct otherwise
uninitialised. correct_for_gaps then blindly read td.segments.back().gexon.exon_id,
which is undefined behaviour — on most platforms the stack holds 0, so
transcripts with high exon IDs (e.g. CHS.42179.1 at exon_id=16) were
spuriously eliminated while nearby isoforms with low exon IDs passed.

Fix (src/evaluate.cpp): walk backwards through td.segments to find the
last segment that has has_gexon==true before comparing exon IDs.

## Bug 2: Rust get_strands_to_check respects splice-strand tags for long reads

The C++ evaluator always checks both '+' and '-' strands for long reads,
regardless of any XS/ts tag. Rust was only checking both strands for
single-exon long reads; multi-exon reads with a strand tag checked only
that one strand, missing transcripts on the opposite strand.

Fix (bramble-rs/src/evaluate.rs): unconditionally return {'+','-'} for
all long reads.

## Bug 3: uint8_t overflow in g2t getTreeForStrand truncates refid

getTreeForStrand, getGuideExons, getGuideExonForTid and indexTrees all
accepted the reference-sequence ID as uint8_t. When a BAM contains more
than ~230 non-annotation reference sequences (alt contigs, patches, etc.)
that appear in reads before the target contig, the gseq_id assigned by
GffNameList exceeds 255. Casting to uint8_t wraps the ID modulo 256,
causing the evaluator to query the wrong chromosome's interval tree and
produce spurious cross-chromosome transcript matches (e.g. a chr17_alt
alignment incorrectly matched to CHS.18641.13 on chr16).

Fix (include/g2t.h, src/g2t.cpp): change all uint8_t refid parameters
to int; add refid < 0 guard in getTreeForStrand.

Long-read golden regenerated (2,475,270 records); Rust and C++ agree
exactly.

+++ include/g2t.h
@@ -85,7 +85,7 @@ namespace bramble {
 
   private:
   
-    IntervalTree *getTreeForStrand(uint8_t refid, char strand);
+    IntervalTree *getTreeForStrand(int refid, char strand);
 
   public:
     // If tid_name is already part of the tree, then return the existing
@@ -98,20 +98,20 @@ namespace bramble {
     void createTree(int refid);
 
     // Add guide exon with TID and transcript start
-    void addInterval(refid_t refid, const tid_t &tid, 
+    void addInterval(refid_t refid, const tid_t &tid,
                     IntervalData interval, char strand,
                     const char* ref_name);
 
     // Index tree after guides have been added
-    void indexTrees(uint8_t refid);
+    void indexTrees(int refid);
 
     bool
-    getGuideExonForTid(uint8_t refid, char strand, tid_t tid, 
+    getGuideExonForTid(int refid, char strand, tid_t tid,
                       uint32_t start, uint32_t end, GuideExon &gexon);
 
     // Find all guide exons that overlap with a read exon
     bool
-    getGuideExons(uint8_t refid, char strand, GSeg exon, 
+    getGuideExons(int refid, char strand, GSeg exon,
                  ReadEvaluationConfig config, ExonStatus status,
                  std::vector<GuideExon> &gexons);
